Make receipt a separate component
