+++ src/components/Expense/ExpensePage.tsx
@@ -2,17 +2,16 @@ import { type Expense, type ExpenseParticipant, type User } from '@prisma/client
 import { isSameDay } from 'date-fns';
 import { Banknote } from 'lucide-react';
 import { type User as NextUser } from 'next-auth';
-import Image from 'next/image';
-import React, { useMemo } from 'react';
 
 // import { api } from '~/utils/api';
 import { toUIString } from '~/utils/numbers';
 
+import type { FC } from 'react';
 import { displayName, toUIDate } from '~/utils/strings';
 import { UserAvatar } from '../ui/avatar';
 import { CategoryIcons } from '../ui/categoryIcons';
-import { AppDrawer } from '../ui/drawer';
 import { Separator } from '../ui/separator';
+import { Receipt } from './Receipt';
 
 interface ExpenseDetailsProps {
   user: NextUser;
@@ -26,24 +25,10 @@ interface ExpenseDetailsProps {
   storagePublicUrl?: string;
 }
 
-const ExpenseDetails: React.FC<ExpenseDetailsProps> = ({ user, expense, storagePublicUrl }) => {
+const ExpenseDetails: FC<ExpenseDetailsProps> = ({ user, expense, storagePublicUrl }) => {
   const CategoryIcon = CategoryIcons[expense.category] ?? Banknote;
 
   // const sendNotificationMutation = api.user.sendExpensePushNotification.useMutation();
-
-  const receiptThumbnail = useMemo(() => {
-    return (
-      <Image
-        src={`${storagePublicUrl}/${expense.fileKey}`}
-        alt="Expense receipt"
-        width={56}
-        height={56}
-        data-loaded="false"
-        onLoad={setDataLoaded}
-        className="h-14 w-14 rounded-md object-cover object-center data-[loaded=false]:animate-pulse data-[loaded=false]:bg-gray-100/10"
-      />
-    );
-  }, [expense.fileKey, storagePublicUrl]);
 
   return (
     <div className="">
@@ -83,24 +68,7 @@ const ExpenseDetails: React.FC<ExpenseDetailsProps> = ({ user, expense, storageP
         </div>
         <div>
           {expense.fileKey ? (
-            <AppDrawer
-              trigger={receiptThumbnail}
-              leftAction="Close"
-              title="Expense Receipt"
-              className="h-[98vh]"
-            >
-              <div className="mb-8 overflow-scroll">
-                <Image
-                  src={`${storagePublicUrl}/${expense.fileKey}`}
-                  width={300}
-                  height={800}
-                  alt="Expense receipt"
-                  data-loaded="false"
-                  onLoad={setDataLoaded}
-                  className="h-full w-full rounded-2xl object-cover data-[loaded=false]:animate-pulse data-[loaded=false]:bg-gray-100/10"
-                />
-              </div>
-            </AppDrawer>
+            <Receipt fileKey={expense.fileKey} url={storagePublicUrl ?? ''} />
           ) : null}
         </div>
       </div>
@@ -138,10 +106,6 @@ const ExpenseDetails: React.FC<ExpenseDetailsProps> = ({ user, expense, storageP
       </div>
     </div>
   );
-};
-
-const setDataLoaded = (event: React.SyntheticEvent<HTMLImageElement>) => {
-  event.currentTarget.setAttribute('data-loaded', 'true');
 };
 
 export default ExpenseDetails;
